# api.yml
openapi: "3.0.0"
info:
    version: 1.0.0
    title: Checkout Backend Service - IFP 2026
    description: |
        Backend service for e-commerce checkout system with dynamic promotions.
        Supports product management, shopping cart, promotions, and order processing.
    license:
        name: MIT
    contact:
        name: API Support
        email: support@example.com
servers:
    - url: http://localhost:8080/api
      description: Local development server
    - url: https://api.example.com/api
      description: Production server
...
paths:
...
    /products:
        get:
            tags:
                - Products
            summary: Get list of products
            description: Retrieve products with pagination and filtering capabilities
            operationId: getProducts
            parameters:
                - name: page
                  in: query
                  required: false
                  schema:
                      type: integer
                      minimum: 1
                      default: 1
                  description: Page number for pagination
                - name: limit
                  in: query
                  required: false
                  schema:
                      type: integer
                      minimum: 1
                      maximum: 100
                      default: 10
                  description: Number of items per page
                - name: search
                  in: query
                  required: false
                  schema:
                      type: string
                  description: Search by product name or SKU
                - name: sku
                  in: query
                  required: false
                  schema:
                      type: string
                  description: Filter by exact SKU
                - name: min_price
                  in: query
                  required: false
                  schema:
                      type: number
                      format: float
                      minimum: 0
                  description: Minimum price filter
                - name: max_price
                  in: query
                  required: false
                  schema:
                      type: number
                      format: float
                      minimum: 0
                  description: Maximum price filter
            responses:
                "200":
                    description: Products retrieved successfully
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ProductListResponse"
                "400":
                    description: Invalid request parameters
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /products/{sku}:
        get:
            tags:
                - Products
            summary: Get product by SKU
            description: Retrieve detailed product information by SKU
            operationId: getProductBySKU
            parameters:
                - name: sku
                  in: path
                  required: true
                  schema:
                      type: string
                  description: Product SKU
            responses:
                "200":
                    description: Product retrieved successfully
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ProductResponse"
                "404":
                    description: Product not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

    # ============================================
    # CART ENDPOINTS
    # ============================================
...
    /orders:
        get:
            tags:
                - Orders
            summary: Get list of orders
            description: Retrieve orders with pagination and filtering
            operationId: getOrders
            parameters:
                - name: page
                  in: query
                  required: false
                  schema:
                      type: integer
                      minimum: 1
                      default: 1
                  description: Page number for pagination
                - name: limit
                  in: query
                  required: false
                  schema:
                      type: integer
                      minimum: 1
                      maximum: 100
                      default: 10
                  description: Number of items per page
                - name: status
                  in: query
                  required: false
                  schema:
                      type: string
                      enum:
                          - pending_payment
                          - paid
                          - processing
                          - shipped
                          - delivered
                          - completed
                          - cancelled
                          - payment_failed
                  description: Filter by order status
                - name: start_date
                  in: query
                  required: false
                  schema:
                      type: string
                      format: date-time
                  description: Filter orders created after this date
                - name: end_date
                  in: query
                  required: false
                  schema:
                      type: string
                      format: date-time
                  description: Filter orders created before this date
            responses:
                "200":
                    description: Orders retrieved successfully
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/OrderListResponse"
                "400":
                    description: Invalid request parameters
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /orders/{id}:
        get:
            tags:
                - Orders
            summary: Get order details
            description: Retrieve detailed order information by ID
            operationId: getOrderByID
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
                  description: Order ID
            responses:
                "200":
                    description: Order details retrieved successfully
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/OrderDetailResponse"
                "404":
                    description: Order not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /orders/{id}/status:
        put:
            tags:
                - Orders
            summary: Update order status
            description: Update the status of an order (for order tracking/fulfillment)
            operationId: updateOrderStatus
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
                  description: Order ID
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateOrderStatusRequest"
            responses:
                "200":
                    description: Order status updated successfully
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/UpdateOrderStatusResponse"
                "400":
                    description: Invalid request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Order not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "422":
                    description: Invalid status transition
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
components:
    schemas:
...
        ProductResponse:
            type: object
            properties:
                sku:
                    type: string
                    example: "120P90"
                name:
                    type: string
                    example: "Google Home"
                price:
                    type: number
                    format: float
                    example: 49.99
                stock_quantity:
                    type: integer
                    example: 1043
                created_at:
                    type: string
                    format: date-time
                updated_at:
                    type: string
                    format: date-time
        ProductListResponse:
            type: object
            properties:
                products:
                    type: array
                    items:
                        $ref: "#/components/schemas/ProductResponse"
                pagination:
                    $ref: "#/components/schemas/Pagination"

        # ==========================================
        # CART SCHEMAS
        # ==========================================
...
        OrderItem:
            type: object
            properties:
                sku:
                    type: string
                    example: "120P90"
                quantity:
                    type: integer
                    example: 3
                unit_price:
                    type: number
                    format: float
                    example: 49.99
                original_subtotal:
                    type: number
                    format: float
                    example: 149.97
                discount:
                    type: number
                    format: float
                    example: 49.99
                final_price:
                    type: number
                    format: float
                    example: 99.98
                promotion_applied:
                    type: string
                    example: "Buy 3 Google Homes for price of 2"
        PromotionBreakdown:
            type: object
            properties:
                promotion:
                    type: string
                    description: Promotion name
                    example: "Buy 3 Google Homes for price of 2"
                saving:
                    type: number
                    format: float
                    description: Total saving from this promotion
                    example: 49.99
                items_affected:
                    type: array
                    items:
                        type: string
                    description: SKUs affected by this promotion
                    example: ["120P90"]
        CheckoutSummary:
            type: object
            properties:
                subtotal:
                    type: number
                    format: float
                    example: 5579.96
                total_discount:
                    type: number
                    format: float
                    example: 79.99
                grand_total:
                    type: number
                    format: float
                    example: 5499.97
                promotion_breakdown:
                    type: array
                    items:
                        $ref: "#/components/schemas/PromotionBreakdown"
        PaymentInfo:
            type: object
            properties:
                payment_id:
                    type: string
                    format: uuid
                    example: "880e8400-e29b-41d4-a716-446655440000"
                status:
                    type: string
                    enum: [pending, success, failed]
                    example: "pending"
                amount:
                    type: number
                    format: float
                    example: 5499.97
                payment_method:
                    type: string
                    example: "credit_card"
                transaction_id:
                    type: string
                    example: "txn_12345678"
...
        OrderListItem:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                order_number:
                    type: string
                    example: "ORD-20260128-0001"
                user_id:
                    type: string
                    format: uuid
                grand_total:
                    type: number
                    format: float
                    example: 5499.97
                status:
                    type: string
                    example: "paid"
                items_count:
                    type: integer
                    example: 3
                created_at:
                    type: string
                    format: date-time
        OrderListResponse:
            type: object
            properties:
                orders:
                    type: array
                    items:
                        $ref: "#/components/schemas/OrderListItem"
                pagination:
                    $ref: "#/components/schemas/Pagination"

        # ==========================================
        # ORDER DETAIL SCHEMAS
        # ==========================================
        StatusHistory:
            type: object
            properties:
                status:
                    type: string
                    example: "pending_payment"
                timestamp:
                    type: string
                    format: date-time
        OrderDetailResponse:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                order_number:
                    type: string
                    example: "ORD-20260128-0001"
                user_id:
                    type: string
                    format: uuid
                items:
                    type: array
                    items:
                        $ref: "#/components/schemas/OrderItem"
                summary:
                    $ref: "#/components/schemas/CheckoutSummary"
                payment:
                    $ref: "#/components/schemas/PaymentInfo"
                shipping_address:
                    type: string
                    example: "Jl. Contoh No. 123, Jakarta, 12345, Indonesia"
                status:
                    type: string
                    example: "paid"
                status_history:
                    type: array
                    items:
                        $ref: "#/components/schemas/StatusHistory"
                created_at:
                    type: string
                    format: date-time
                updated_at:
                    type: string
                    format: date-time

        # ==========================================
        # ORDER STATUS UPDATE SCHEMAS
        # ==========================================
        UpdateOrderStatusRequest:
            type: object
            required:
                - status
            properties:
                status:
                    type: string
                    enum:
                        - pending_payment
                        - paid
                        - processing
                        - shipped
                        - delivered
                        - completed
                        - cancelled
                        - payment_failed
                    description: New order status
                    example: "shipped"
                note:
                    type: string
                    description: Optional note about the status update
                    example: "Package shipped via JNE with tracking number JNE123456"
        UpdateOrderStatusResponse:
            type: object
            properties:
                order_id:
                    type: string
                    format: uuid
                status:
                    type: string
                    example: "shipped"
                message:
                    type: string
                    example: "Order status updated successfully"

        # ==========================================
        # COMMON SCHEMAS
        # ==========================================
        Pagination:
            type: object
            properties:
                current_page:
                    type: integer
                    example: 1
                total_pages:
                    type: integer
                    example: 10
                total_items:
                    type: integer
                    example: 100
                items_per_page:
                    type: integer
                    example: 10
        ErrorResponse:
            type: object
            properties:
                status:
                    type: string
                    example: "error"
                message:
                    type: string
                    example: "Invalid request parameters"
    securitySchemes:
        ApiKeyAuth:
            type: apiKey
            in: header
            name: X-API-Key

# Optional: Add security if needed
# security:
#   - ApiKeyAuth: []